Endpoint tests for user registration and user login completed. Refractored file structure because I needed to import app for testing.
