+++ requirements.txt
@@ -8,25 +8,40 @@ click==8.1.7
 colorama==0.4.6
 cryptography==43.0.3
 dnspython==2.6.1
-flask==3.0.3
+Flask==3.0.3
+Flask-Cors==5.0.0
 Flask-JWT-Extended==4.6.0
 Flask-SocketIO==5.4.1
 h11==0.14.0
 idna==3.10
-importlib-metadata==8.5.0
+importlib_metadata==8.5.0
+iniconfig==2.0.0
 itsdangerous==2.2.0
-jinja2==3.1.4
+Jinja2==3.1.4
+jwt==1.3.1
 MarkupSafe==2.1.5
+mirakuru==2.5.3
 mongoengine==0.29.1
+mongomock==4.2.0.post1
+packaging==24.1
+pluggy==1.5.0
+port-for==0.7.4
+psutil==6.1.0
 pycparser==2.22
 PyJWT==2.9.0
 pymongo==4.10.1
+pytest==8.3.3
+pytest-mongo==3.1.0
+pytest-mongodb==2.4.0
 python-dotenv==1.0.1
 python-engineio==4.10.1
 python-socketio==5.11.4
+pytz==2024.2
+PyYAML==6.0.2
 requests==2.32.3
+sentinels==1.0.0
 simple-websocket==1.1.0
 urllib3==2.2.3
-werkzeug==3.0.4
+Werkzeug==3.0.4
 wsproto==1.2.0
 zipp==3.20.2
