--- socket_manager.py
@@ -1,3 +1,0 @@
-from flask_socketio import SocketIO
-
-socketio = SocketIO()
