+++ src/socket_manager.py
@@ -1,0 +1,3 @@
+from flask_socketio import SocketIO
+
+socketio = SocketIO()
