--- tests/conftest.py
@@ -1,2 +1,0 @@
-import pytest
-from app import app as flask_app
